+++ src/index.html
@@ -36,6 +36,7 @@
     <include src="./partials/burgerMenu.html"></include>
     <include src="./partials/readMoreModal.html"></include>
     <include src="./partials/buyNowModal.html"></include>
+    <include src="./partials/productCardModal.html"></include>
     <include src="./partials/productInfoModal.html"></include>
     <include src="./partials/franchiseModal.html"></include>
     <include src="./partials/ourLocationModal.html"></include>
