+++ src/js/header-scroll.js
@@ -1,0 +1,13 @@
+window.onscroll = () => changeHeaderBackground();
+
+function changeHeaderBackground() {
+  const header = document.querySelector('[data-header]');
+  const headerOffsetTrigger = header.offsetHeight;
+  const pageOffset = window.pageYOffset;
+
+  if (pageOffset > headerOffsetTrigger) {
+    header.classList.add('header--transparent');
+  } else {
+    header.classList.remove('header--transparent');
+  }
+}
