+++ src/partials/burgerMenu.html
@@ -40,7 +40,7 @@
     <button
       class="button buy-now-button buy-now-button--tablet-none"
       type="button"
-      buy-now-modal-open2
+      product-card-modal-open
     >
       Buy now
     </button>
