+++ src/partials/productCardModal.html
@@ -1,6 +1,6 @@
-<div class="backdrop is-hidden" buy-now-modal>
+<div class="backdrop is-hidden" product-card-modal>
   <div class="modal_buy-now">
-    <button class="modal_buy-now-close" type="button" buy-now-modal-close>
+    <button class="modal_buy-now-close" type="button" product-card-modal-close>
       <svg class="icon">
         <use href="./images/sprite.svg#ep_close"></use>
       </svg>
