+++ src/sass/layout/_buyNowModal.scss
@@ -92,10 +92,6 @@
         color: $secondary-text-color;
       }
 
-      &:invalid {
-        border: 1px solid #e78b8b;
-      }
-
       @media screen and (max-width: 767px) {
         margin-bottom: 1px;
       }
