+++ src/sass/layout/_productInfoModal.scss
@@ -175,25 +175,40 @@
     width: 100%;
     height: 42px;
 
+    outline: transparent;
+    color: $secondary-text-color;
     background: $input-bg-color;
     border: 1px solid $bg-color;
     border-radius: 10px;
+
+    transition: border 250ms linear;
+  }
+
+  &__form-input:hover,
+  &__form-input:focus,
+  &__form-comment:hover,
+  &__form-comment:focus {
+    border: 1px solid $accent-color;
   }
 
   &__form-input::placeholder,
   &__form-comment::placeholder {
     font-size: 14px;
     line-height: calc(19.6 / 14);
+    color: $secondary-text-color;
   }
 
   &__form-comment {
     width: 100%;
     height: 84px;
     padding: 11px 16px;
-    background: rgba(248, 210, 78, 0.1);
+    background-color: $input-bg-color;
     border: 1px solid $bg-color;
     border-radius: 10px;
     resize: none;
+    outline: transparent;
+    color: $secondary-text-color;
+    transition: border 250ms linear;
   }
 
   &__btn {
